requirements.txt
flask==3.0.0
openai==1.0.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
flask-cors==5.0.0
pyngrok==7.0.5
beautifulsoup4==4.12.2
requests==2.31.0
